+++ app/src/shared/components/MobileNav/NavItem/navItem.pcss
