+++ app/src/shared/components/Nav/AvatarItem/avatarItem.pcss
