+++ app/src/shared/components/Nav/AvatarItem/index.jsx
