+++ app/src/shared/components/Nav/index.jsx
@@ -1,116 +1,124 @@
 // @flow
 
-import * as React from 'react'
-import classNames from 'classnames'
+import React from 'react'
+import cx from 'classnames'
+import { withRouter, type Location } from 'react-router-dom'
+import { compose } from 'redux'
+import { connect } from 'react-redux'
+import { I18n, Translate } from 'react-redux-i18n'
+import { selectUserData } from '$shared/modules/user/selectors'
+import type { User } from '$shared/flowtype/user-types'
+import Link from '$shared/components/Link'
+import LogoItem from './LogoItem'
+import DropdownItem from './DropdownItem'
+import LinkItem from './LinkItem'
+import AvatarItem from './AvatarItem'
+import routes from '$routes'
+import navigationLinks from '$docs/components/DocsLayout/Navigation/navLinks'
+
 import styles from './nav.pcss'
-import NavLogo from './NavLogo'
-import NavHamburger from './NavHamburger'
-import NavItem from './NavItem'
 
-type Props = {
-    children: React.Node,
-    expand?: boolean,
-    opaque?: boolean,
-    overlay?: boolean,
-    label?: string,
-    className?: string,
+type StateProps = {
+    currentUser: ?User,
+}
+type Props = StateProps & {
+    location: Location,
 }
 
-type State = {
-    open: boolean,
-}
+const mapStateToProps = (state): StateProps => ({
+    currentUser: selectUserData(state),
+})
 
-const toggleBodyClass = (className, state) => {
-    if (document.body) {
-        const { classList } = document.body
-        if (state) {
-            classList.add(className)
-        } else {
-            classList.remove(className)
-        }
-    }
-}
+const Nav = compose(
+    connect(mapStateToProps),
+    withRouter,
+)(({ currentUser, location }: Props) => (
+    <nav
+        className={styles.root}
+    >
+        <div>
+            <LogoItem />
+        </div>
+        <div>
+            <DropdownItem label={I18n.t('general.core')} to="#" align="right">
+                <Link
+                    className={Nav.styles.link}
+                    to={routes.streams()}
+                >
+                    <Translate value="general.streams" />
+                </Link>
+                <Link
+                    className={Nav.styles.link}
+                    to={routes.canvases()}
+                >
+                    <Translate value="general.canvases" />
+                </Link>
+                <Link
+                    className={Nav.styles.link}
+                    to={routes.dashboards()}
+                >
+                    <Translate value="general.dashboards" />
+                </Link>
+                <Link
+                    className={Nav.styles.link}
+                    to={routes.products()}
+                >
+                    <Translate value="general.products" />
+                </Link>
+                <Link
+                    className={Nav.styles.link}
+                    to={routes.purchases()}
+                >
+                    <Translate value="general.purchases" />
+                </Link>
+                <Link
+                    className={Nav.styles.link}
+                    to={routes.transactions()}
+                >
+                    <Translate value="general.transactions" />
+                </Link>
+            </DropdownItem>
+            <LinkItem to={routes.root()}>
+                <Translate value="general.marketplace" />
+            </LinkItem>
+            <DropdownItem label={I18n.t('general.docs')} to="#" align="left">
+                {Object.keys(navigationLinks).map((key) => (
+                    <Link
+                        key={key}
+                        to={navigationLinks[key]}
+                        target="_blank"
+                        rel="noopener noreferrer"
+                        className={Nav.styles.link}
+                    >
+                        {key}
+                    </Link>
+                ))}
+            </DropdownItem>
+            {!!currentUser && (
+                <AvatarItem user={currentUser} />
+            )}
+            {!currentUser && (
+                <LinkItem
+                    to={routes.login({
+                        redirect: location.pathname,
+                    })}
+                    className={Nav.styles.link}
+                >
+                    <Translate value="general.signIn" />
+                </LinkItem>
+            )}
+            {!currentUser && (
+                <LinkItem
+                    to={routes.signUp()}
+                    className={cx(Nav.styles.link, Nav.styles.button)}
+                >
+                    <Translate value="general.signUp" />
+                </LinkItem>
+            )}
+        </div>
+    </nav>
+))
 
-class Nav extends React.Component<Props, State> {
-    state = {
-        open: false,
-    }
-
-    onToggleClick = (e: SyntheticInputEvent<EventTarget>) => {
-        const open = !this.state.open
-
-        e.preventDefault()
-        this.setState({
-            open,
-        })
-        toggleBodyClass('no-scroll', open)
-    }
-
-    onClose = () => {
-        this.setState({
-            open: false,
-        })
-        toggleBodyClass('no-scroll', false)
-    }
-
-    render() {
-        const { open } = this.state
-        const {
-            children,
-            expand,
-            opaque,
-            overlay,
-            label,
-            className,
-        } = this.props
-
-        return (
-            <nav
-                className={classNames(styles.nav, className, {
-                    [styles.opaque]: opaque,
-                    [styles.open]: open,
-                    [styles.fullWidth]: expand,
-                    [styles.overlay]: overlay,
-                })}
-            >
-                <div className="container">
-                    <div className={styles.inner}>
-                        <NavLogo />
-                        {!!label && (
-                            <div className={classNames(styles.marketplaceLabel, 'd-md-none')}>
-                                {label}
-                            </div>
-                        )}
-                        <NavHamburger onClick={this.onToggleClick} opaqueNav={opaque} />
-                        <div className={styles.navItemsWrapper}>
-                            <div className={styles.toggleWrapper}>
-                                <div className="container">
-                                    <NavHamburger open onClick={this.onToggleClick} />
-                                </div>
-                            </div>
-                            <ul className={styles.navItems}>
-                                {React.Children.map(children, (child) => child && (
-                                    <NavItem opaqueNav={opaque}>
-                                        {React.cloneElement(child, {
-                                            closeNav: this.onClose,
-                                        })}
-                                    </NavItem>
-                                ))}
-                            </ul>
-                        </div>
-                    </div>
-                </div>
-            </nav>
-        )
-    }
-}
-
-export { default as NavDivider } from './NavDivider'
-export { default as NavDropdown } from './NavDropdown'
-export { default as NavHamburger } from './NavHamburger'
-export { default as NavItem } from './NavItem'
-export { default as NavLink } from './NavLink'
-export { default as NavLogo } from './NavLogo'
-export { default as NavLabel } from './NavLabel'
+Nav.styles = styles
 
 export default Nav
