--- app/src/shared/components/Nav2/nav.pcss
@@ -1,47 +1,0 @@
-.root {
-  display: flex;
-  justify-content: space-between;
-  padding: var(--um);
-}
-
-.root > div {
-  align-items: center;
-  display: flex;
-  height: calc(2 * var(--um));
-}
-
-.root > div > * + * {
-  margin-left: calc(1.5 * var(--um));
-}
-
-.link {
-  display: block;
-  font-family: var(--mono);
-  font-size: calc(0.75 * var(--um));
-  font-weight: var(--medium);
-  letter-spacing: calc(0.125 * var(--um));
-  text-transform: uppercase;
-  transition: 200ms ease-in-out;
-  transition-property: background-color, border-color, color;
-
-  &:--enter {
-    transition-duration: 50ms;
-  }
-}
-
-.button {
-  border: 1px solid #A3A3A3;
-  border-radius: calc(0.125 * var(--um));
-  line-height: calc(1.375 * var(--um));
-  padding: 0 calc(0.75 * var(--um));
-
-  &:--enter {
-    border-color: #323232;
-  }
-}
-
-@media (--md-up) {
-  .root {
-    padding: calc(1.5 * var(--um));
-  }
-}
